\name{cor_scores_genes}
\docType{data}
\alias{cor_scores_genes}
\title{
    Genes used to define correlated scores of immune response
}
\arguments{
    \item{metadata}{Logical, whether only metadata should be returned}
}
\description{
   Genes used to define correlated scores of immune response.
}
\usage{
    cor_scores_genes(metadata = FALSE)
}
\examples{
    cor_scores_genes(metadata = FALSE)
}
\format{Character}
\value{Returns a \code{character} vector.}
\author{
    Oscar Lapuente-Santana
}
\references{
    Lapuente-Santana, Óscar, Maisa van Genderen, Peter A. J. Hilbers, Francesca Finotello, and Federica Eduati.
    2021. “Interpretable Systems Biomarkers Predict Response to Immune-Checkpoint Inhibitors.” Patterns, 100293.
    https://doi.org/10.1016/j.patter.2021.100293.
}
