% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData-pkg.R
\docType{package}
\name{easierData-pkg}
\alias{easierData}
\alias{easierData-package}
\alias{easierData-pkg}
\title{easierData: easier data package}
\description{
This experiment data package provides access to an exemplary
bladder cancer dataset with both processed RNA-seq data and
information on response to ICB therapy publicly available via
IMvigor210CoreBiologies package, and to internal data
indispensable for the functional performance of easier package.
}
\references{
Óscar Lapuente-Santana, Maisa van Genderen, Peter A. J. Hilbers,
Francesca Finotello, and Federica Eduati. 2021. Interpretable Systems
Biomarkers Predict Response to Immune-Checkpoint Inhibitors.
Patterns, 100293. https://doi.org/10.1016/j.patter.2021.100293.
}
\author{
\strong{Maintainer}: Oscar Lapuente-Santana \email{o.lapuente.santana@tue.nl} (\href{https://orcid.org/0000-0003-1995-8393}{ORCID})

Authors:
\itemize{
  \item Federico Marini \email{marinif@uni-mainz.de} (\href{https://orcid.org/0000-0003-3252-7758}{ORCID})
  \item Arsenij Ustjanzew \email{arsenij.ustjanzew@uni-mainz.de} (\href{https://orcid.org/0000-0002-1014-4521}{ORCID})
  \item Francesca Finotello \email{francesca.finotello@i-med.ac.at} (\href{https://orcid.org/0000-0003-0712-4658}{ORCID})
  \item Federica Eduati \email{f.eduati@tue.nl} (\href{https://orcid.org/0000-0002-7822-3867}{ORCID})
}

}
\keyword{internal}
