% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_cor_scores_genes}
\alias{get_cor_scores_genes}
\title{Title}
\usage{
get_cor_scores_genes()
}
\value{
A character vector TODO
}
\description{
Genes used to define correlated scores of immune response
}
\examples{
if (interactive()) {
  # TODO
}
}
\references{
Óscar Lapuente-Santana, Maisa van Genderen, Peter A. J. Hilbers,
Francesca Finotello, and Federica Eduati. 2021. Interpretable Systems Biomarkers
Predict Response to Immune-Checkpoint Inhibitors. Patterns, 100293.
https://doi.org/10.1016/j.patter.2021.100293.

Mariathasan, S., Turley, S., Nickles, D. et al. TGFβ attenuates
tumour response to PD-L1 blockade by contributing to exclusion of T cells.
Nature 554, 544–548 (2018). https://doi.org/10.1038/nature25501.
}
