% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_group_lrpairs}
\alias{get_group_lrpairs}
\title{Title}
\usage{
get_group_lrpairs()
}
\value{
TODO
}
\description{
Groups of ligand-receptor pairs
}
\details{
Information to group ligand-receptor pairs because of sharing the same gene,
either as ligand or receptor.
}
\examples{
if (interactive()) {
  # TODO
}
}
