% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_HGNC_annotation}
\alias{get_HGNC_annotation}
\title{Title}
\source{
https://www.genenames.org/tools/multi-symbol-checker/
}
\usage{
get_HGNC_annotation()
}
\value{
TODO
}
\description{
Gene symbols approved annotations.
}
\examples{
if (interactive()) {
  # TODO
}
}
\references{
Tweedie S, Braschi B, Gray KA, Jones TEM, Seal RL, Yates B, Bruford EA.
Genenames.org: the HGNC and VGNC resources in 2021. Nucleic Acids Res.
PMID: 33152070 PMCID: PMC7779007 DOI: 10.1093/nar/gkaa980
}
