% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_lr_frequency_TCGA}
\alias{get_lr_frequency_TCGA}
\title{Title}
\usage{
get_lr_frequency_TCGA()
}
\value{
TODO
}
\description{
Frequency of each ligand-receptor pair feature across the whole TCGA database
}
\examples{
if (interactive()) {
  # TODO
}
}
\references{
The Cancer Genome Atlas Research Network., Weinstein, J., Collisson, E.
et al. The Cancer Genome Atlas Pan-Cancer analysis project. Nat Genet 45, 1113–1120 (2013).
https://doi.org/10.1038/ng.2764
}
