% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_opt_models}
\alias{get_opt_models}
\title{Title}
\usage{
get_opt_models()
}
\value{
TODO
}
\description{
Cancer-specifc model feature parameters
}
\details{
Cancer-specifc model feature parameters learned, for each quantitative descriptor
during training using multi-task learning with randomized cross-validation repeated 100 times.
For each quantitative descriptor, 1000 models are available (100 per task).
}
\examples{
if (interactive()) {
  # TODO
}
}
