% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_scores_signature_genes}
\alias{get_scores_signature_genes}
\title{Title}
\usage{
get_scores_signature_genes()
}
\value{
TODO
}
\description{
Title
}
\examples{
if (interactive()) {
  # TODO
}
}
