% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_TCGA_mean_pancancer}
\alias{get_TCGA_mean_pancancer}
\title{Get the mean of the TPM expression of each gene across all TCGA cancer types}
\usage{
get_TCGA_mean_pancancer()
}
\value{
A numeric vector with gene TPM mean values.
}
\description{
Obtain the mean of the TPM expression of each gene across all TCGA cancer types.
It is required for normalization of input TPM gene expression data.
}
\examples{
if (interactive()) {
  TCGA_mean_pancancer <- get_TCGA_mean_pancancer()
}
}
\references{
The Cancer Genome Atlas Research Network., Weinstein, J., Collisson, E.
et al. The Cancer Genome Atlas Pan-Cancer analysis project. Nat Genet 45, 1113–1120 (2013).
https://doi.org/10.1038/ng.2764
}
