% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{list_easierData}
\alias{list_easierData}
\title{List the datasets}
\usage{
list_easierData()
}
\value{
A data.frame for a compact visualization
}
\description{
List the datasets contained in the easierData package
}
\examples{
df <- list_easierData()
df
}
