% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData-pkg.R
\docType{package}
\name{easierData-pkg}
\alias{easierData-pkg}
\title{easierData}
\description{
This ExperimentHub package provides access to an exemplary
bladder cancer dataset with both processed RNA-seq data and
information on response to ICB therapy publicly available via
IMvigor210CoreBiologies package, and to internal data
indispensable for the functional performance of easier package.
}
\references{
Óscar Lapuente-Santana, Maisa van Genderen, Peter A. J. Hilbers,
Francesca Finotello, and Federica Eduati. 2021. Interpretable Systems
Biomarkers Predict Response to Immune-Checkpoint Inhibitors.
Patterns, 100293. https://doi.org/10.1016/j.patter.2021.100293.
}
