\name{easierData}
\docType{package}
\alias{easierData}
\title{
    easierData
}
\description{
    This experimentHub package provides access to an exemplary bladder cancer dataset with both processed RNA-seq       data and information on response to ICB therapy publicly available via IMvigor210CoreBiologies package, and to      internal data indispensable for the functional performance of easier package.
}
\author{
    Oscar Lapuente-Santana
}
\references{
    Mariathasan, S., Turley, S., Nickles, D. et al. TGFβ attenuates tumour response to PD-L1 blockade by
    contributing to exclusion of T cells. Nature 554, 544–548 (2018). https://doi.org/10.1038/nature25501.

    Lapuente-Santana, Óscar, Maisa van Genderen, Peter A. J. Hilbers, Francesca Finotello, and Federica Eduati.
    2021. “Interpretable Systems Biomarkers Predict Response to Immune-Checkpoint Inhibitors.” Patterns, 100293.
    https://doi.org/10.1016/j.patter.2021.100293.
}
