% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_cor_scores_genes}
\alias{get_cor_scores_genes}
\title{Get the list of genes used to define correlated
scores of immune response}
\usage{
get_cor_scores_genes()
}
\value{
A character vector with gene names.
}
\description{
Obtain the list of genes used to define each of the correlated
scores of immune response. These scores were found to be highly
correlated across all 18 cancer types.
}
\examples{
if (interactive()) {
    cor_scores_genes <- get_cor_scores_genes()
}
}
\references{
Óscar Lapuente-Santana, Maisa van Genderen, Peter A. J. Hilbers,
Francesca Finotello, and Federica Eduati. 2021. Interpretable Systems
Biomarkers Predict Response to Immune-Checkpoint Inhibitors.
Patterns, 100293. https://doi.org/10.1016/j.patter.2021.100293.

Mariathasan, S., Turley, S., Nickles, D. et al. TGFB attenuates
tumour response to PD-L1 blockade by contributing to exclusion of T cells.
Nature 554, 544–548 (2018). https://doi.org/10.1038/nature25501.
}
