% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_group_lrpairs}
\alias{get_group_lrpairs}
\title{Get the groups of ligand-receptor pairs}
\usage{
get_group_lrpairs()
}
\value{
A list containing grouping information for 56 groups of ligand-receptor pairs.
}
\description{
Obtain the information to group ligand-receptor pairs because of sharing the same gene,
either as ligand or receptor.
}
\examples{
if (interactive()) {
  group_lrpairs <- get_group_lrpairs()
}
}
