% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_Mariathasan2018_PDL1_treatment}
\alias{get_Mariathasan2018_PDL1_treatment}
\title{Get the Mariathasan2018 PDL1 treatment data}
\usage{
get_Mariathasan2018_PDL1_treatment()
}
\value{
A \linkS4class{SummarizedExperiment} object with counts and TPM
quantifications.
}
\description{
Obtain the ...TODO... from Mariathasan et al. (2018)
}
\examples{
if (interactive()) {
  se_mariathasan <- get_Mariathasan2018_PDL1_treatment()
}
}
\references{
Mariathasan et al., full ref TODO
}
