% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easierData_retrieval.R
\name{get_opt_xtrain_stats}
\alias{get_opt_xtrain_stats}
\title{Title}
\usage{
get_opt_xtrain_stats()
}
\value{
TODO
}
\description{
Training set cancer-specific features mean and standard deviation
}
\details{
Cancer-specific features mean and standard deviation of the training set used during
randomized cross-validation repeated 100 times. It is required for normalization of the test set.
}
\examples{
if (interactive()) {
  # TODO
}
}
\references{
Óscar Lapuente-Santana, Maisa van Genderen, Peter A. J. Hilbers,
Francesca Finotello, and Federica Eduati. 2021. Interpretable Systems Biomarkers
Predict Response to Immune-Checkpoint Inhibitors. Patterns, 100293.
https://doi.org/10.1016/j.patter.2021.100293.
}
